+++ src/main/java/wooteco/subway/domain/station/Stations.java
@@ -5,20 +5,24 @@ import wooteco.subway.exception.NoStationExistsException;
 import java.util.List;
 
 public class Stations {
-    private final List<Station> stations;
+	private final List<Station> stations;
 
-    public Stations(List<Station> stations) {
-        this.stations = stations;
-    }
+	private Stations(List<Station> stations) {
+		this.stations = stations;
+	}
 
-    public List<Station> getStations() {
-        return stations;
-    }
+	public static Stations of(List<Station> stations) {
+		return new Stations(stations);
+	}
 
-    public Station extractStationById(Long stationId) {
-        return stations.stream()
-                .filter(it -> it.getId().equals(stationId))
-                .findFirst()
-                .orElseThrow(NoStationExistsException::new);
+	public List<Station> getStations() {
+		return stations;
+	}
+
+	public Station extractStationById(Long stationId) {
+		return stations.stream()
+				.filter(it -> it.getId().equals(stationId))
+				.findFirst()
+				.orElseThrow(NoStationExistsException::new);
     }
 }
