--- src/main/java/wooteco/subway/service/path/GraphService.java
@@ -1,24 +1,0 @@
-package wooteco.subway.service.path;
-
-import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
-import org.jgrapht.graph.DefaultWeightedEdge;
-import org.jgrapht.graph.WeightedMultigraph;
-import org.springframework.stereotype.Service;
-import wooteco.subway.domain.line.Lines;
-import wooteco.subway.domain.path.PathType;
-
-import java.util.List;
-
-@Service
-public class GraphService {
-    public List<Long> findPath(Lines lines, Long source, Long target, PathType type) {
-        WeightedMultigraph<Long, DefaultWeightedEdge> graph
-                = new WeightedMultigraph(DefaultWeightedEdge.class);
-
-        lines.addVertex(graph);
-        lines.setEdge(graph, type);
-
-        DijkstraShortestPath dijkstraShortestPath = new DijkstraShortestPath(graph);
-        return dijkstraShortestPath.getPath(source, target).getVertexList();
-    }
-}
